refactored image service
